+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.61.1](https://github.com/Life-USTC/server/compare/v1.61.0...v1.61.1) (2026-07-17)
+
+
+### Bug Fixes
+
+* **db:** align section lifecycle schema ([#475](https://github.com/Life-USTC/server/issues/475)) ([92f5ba8](https://github.com/Life-USTC/server/commit/92f5ba83c5b759b1715ad5a017646aa78941ed68))
+
+# [1.61.0](https://github.com/Life-USTC/server/compare/v1.60.0...v1.61.0) (2026-07-17)
+
+
+### Features
+
+* **db:** add section lifecycle storage ([#473](https://github.com/Life-USTC/server/issues/473)) ([0187e87](https://github.com/Life-USTC/server/commit/0187e878a1f109e19d6b28ff5a968ea5386f4910))
+* **ops:** add database recovery gate ([#474](https://github.com/Life-USTC/server/issues/474)) ([b8d1030](https://github.com/Life-USTC/server/commit/b8d10309ced9445ef6409bbc1c12ab696a999caa))
+
 # [1.60.0](https://github.com/Life-USTC/server/compare/v1.59.1...v1.60.0) (2026-07-17)